model/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
diffusers>=0.24.0
accelerate>=0.25.0
datasets>=2.14.0
pillow>=10.0.0
numpy>=1.24.0
safetensors>=0.4.0
huggingface-hub>=0.19.0
peft>=0.7.0
bitsandbytes>=0.41.0
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0





